Allow float data to be used with our sparse solver extensions (#391)
* Fixed dtype error in sparse solvers.

* Improved precision of sparse solver grad for float data.

* Added check grad for float data.

+++ theseus/optimizer/sparse_linearization.py
@@ -157,7 +157,7 @@ class SparseLinearization(Linearization):
                 self.num_cols,
                 A_row_ptr,
                 A_col_ind,
-                self.A_val,
-                self.b,
+                self.A_val.double(),
+                self.b.double(),
             ).unsqueeze(2)
-        return self._Atb
+        return self._Atb.to(dtype=self.A_val.dtype)
